feat: Added UI components
- Added UI components for Input, Textarea, and Button elements, and integrated them into the payment link sharing and social media sections.
-  Also added missing packages `@radix-ui/react-slot` and `fortawesome`, which were causing errors.
