+++ paystell-frontend/package.json
@@ -9,11 +9,13 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@fortawesome/free-brands-svg-icons": "^6.7.1",
+    "@fortawesome/react-fontawesome": "^0.2.2",
     "@radix-ui/react-avatar": "^1.1.1",
-    "@radix-ui/react-menubar": "^1.1.2",
-    "@radix-ui/react-popover": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-menubar": "^1.1.2",
+    "@radix-ui/react-popover": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/themes": "^3.1.5",
     "@react-icons/all-files": "^4.1.0",
@@ -23,6 +25,7 @@
     "icons": "^1.0.0",
     "lucide-react": "^0.456.0",
     "next": "14.2.17",
+    "radix-ui": "^1.0.1",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18",
